+++ app/build.gradle
@@ -47,4 +47,5 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
     testImplementation 'androidx.arch.core:core-testing:2.1.0'
+    implementation "androidx.activity:activity-ktx:1.6.1"
 }
