+++ app/src/main/java/ru/netology/nmedia/MainActivity.kt
@@ -2,20 +2,20 @@ package ru.netology.nmedia
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.lifecycle.ViewModelProvider
+import androidx.activity.viewModels
 import ru.netology.nmedia.databinding.ActivityMainBinding
 import ru.netology.nmedia.viewmodel.PostViewModel
+import kotlin.math.roundToInt
 
 
 class MainActivity : AppCompatActivity() {
-    private val binding = ActivityMainBinding.inflate(layoutInflater)
-    private lateinit var viewModel: PostViewModel
-
+    private lateinit var binding: ActivityMainBinding
+    private val viewModel: PostViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        viewModel = ViewModelProvider(this).get(PostViewModel::class.java)
         subscribe()
         setupListener()
     }
@@ -26,20 +26,16 @@ class MainActivity : AppCompatActivity() {
                 author.text = post.author
                 published.text = post.published
                 content.text = post.content
-                likedCount.text = viewModel.countTransform(post.likes)
-                sharedCount.text = viewModel.countTransform(post.shares)
-                viewsCount.text = viewModel.countTransform(post.views)
-                if (post.likedByMe) {
-                    liked.setImageResource(R.drawable.ic_baseline_favorite_24)
-                }
-                if (post.sharedByMe) {
-                    shared.setImageResource(R.drawable.ic_baseline_share_ok_24)
-                }
+                likedCount.text = countTransform(post.likes)
+                sharedCount.text = countTransform(post.shares)
+                viewsCount.text = countTransform(post.views)
+                liked.setImageResource(if (post.likedByMe) R.drawable.ic_baseline_favorite_24 else R.drawable.ic_baseline_favorite_border_24)
+                if (post.sharedByMe) shared.setImageResource(R.drawable.ic_baseline_share_ok_24)
             }
         }
     }
 
-    fun setupListener() {
+    private fun setupListener() {
 
         binding.liked.setOnClickListener {
             viewModel.like()
@@ -49,6 +45,31 @@ class MainActivity : AppCompatActivity() {
             viewModel.share()
         }
 
+    }
+
+    private fun countTransform(x: Int): String {
+        val numberToString = x.toString()
+        val lengthNumber = when (numberToString.length) {
+            in 1..3 -> numberToString
+            in 4..6 -> {
+                val y = x.toDouble() / 1000.0
+                if (y > 1.1 && y < 10) {
+                    (((y * 10.0).roundToInt()) / 10.0).toString() + "K"
+                } else {
+                    y.toInt().toString() + "K"
+                }
+            }
+            in 7..9 -> {
+                val y = x.toDouble() / 1000000.0
+                if (y >= 1.1) {
+                    (((y * 10.0).roundToInt()) / 10.0).toString() + "M"
+                } else {
+                    y.toInt().toString() + "M"
+                }
+            }
+            else -> "не определено"
+        }
+        return lengthNumber
     }
 }
