+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -7,5 +7,4 @@ interface PostRepository {
     fun get(): LiveData<Post>
     fun like()
     fun share()
-    fun countTransform(x: Int): String
 }
