+++ app/src/main/java/ru/netology/nmedia/viewmodel/PostViewModel.kt
@@ -10,5 +10,4 @@ abstract class PostViewModel : ViewModel() {
     val data = repository.get()
     fun like() = repository.like()
     fun share() = repository.share()
-    fun countTransform(x: Int) = repository.countTransform(x)
 }
